requirements.txt
fastapi
uvicorn[standard]
smolagents
huggingface_hub
python-dotenv
duckduckgo-search 